requirements.txt
mediapipe==0.10.14
opencv-python>=4.10.0.84
numpy>=1.26.4
python-dotenv>=1.0.1
pyobjc-framework-Quartz>=10.3
